Conflicts:
	avr/cores/tiny/Arduino.h

+++ avr/cores/tiny/Arduino.h
@@ -534,7 +534,7 @@ unsigned long pulseIn(uint8_t pin, uint8_t state, unsigned long timeout = 100000
 
 #if !defined(tone) && !defined(NO_TONE)
 void initToneTimer(void);
-void tone(uint8_t _pin, unsigned int frequency, unsigned long duration = 0);
+void tone(uint8_t _pin, unsigned long frequency, unsigned long duration = 0);
 void noTone(uint8_t _pin = 255);
 #endif
